Basic support for IPv6 client connections
Replace gethostbyname with getaddrinfo, as the former doesn't support
IPv6.  For now, only use the first returned address, like with
gethostbyname.

Since there is no try-to-connect loop, this will introduce regressions
in some cases.  Example: an IPv4 only capable host that prefers IPv6
(Linux comes to mind) and that attempts to connect to a dual-stacked
domain name.